warnings fixed

+++ modules/core/transport/http/http_chunked_stream.c
@@ -200,7 +200,7 @@ axis2_http_chunked_stream_write (axis2_http_chunked_stream_t *chunked_stream
                   , AXIS2_FAILURE);
       return -1;
    }
-   sprintf(tmp_buf, "%x%s", count, AXIS2_HTTP_CRLF);
+   sprintf(tmp_buf, "%x%s", (unsigned int)count, AXIS2_HTTP_CRLF);
    len = AXIS2_STREAM_WRITE(stream, env, tmp_buf, AXIS2_STRLEN(tmp_buf));
    len = AXIS2_STREAM_WRITE(stream, env, buffer, count);
    AXIS2_STREAM_WRITE(stream, env, AXIS2_HTTP_CRLF, 2);
