+++ modules/core/transport/http/sender/rest_sender.c
@@ -251,7 +251,7 @@ axis2_rest_sender_send(axis2_rest_sender_t *sender,
    if(AXIS2_FALSE == sender_impl->chunked)
    {
       axis2_char_t tmp_buf[10];
-      sprintf(tmp_buf, "%d", strlen(buffer));
+      sprintf(tmp_buf, "%d", (int)strlen(buffer));
       http_header = axis2_http_header_create(env, 
                   AXIS2_HTTP_HEADER_CONTENT_LENGTH, tmp_buf);
       AXIS2_HTTP_SIMPLE_REQUEST_ADD_HEADER(request, env, http_header);
